Feat(Analysis): 학습 현황 조회 로직을 도메인 로직으로 이전
