+++ jabiseo-api/src/main/java/com/jabiseo/analysis/dto/FindTodayLearningResponse.java
@@ -1,22 +1,19 @@
 package com.jabiseo.analysis.dto;
 
+import com.jabiseo.learning.dto.TodayLearningResponse;
+
 public record FindTodayLearningResponse(
     int studyModeCount,
     int studyModeCorrectRate,
     int examModeCount,
     int examModeCorrectRate
 ) {
-    public static FindTodayLearningResponse of(
-        int studyModeCount,
-        int studyModeCorrectRate,
-        int examModeCount,
-        int examModeCorrectRate
-    ) {
+    public static FindTodayLearningResponse from(TodayLearningResponse todayLearningResponse) {
         return new FindTodayLearningResponse(
-            studyModeCount,
-            studyModeCorrectRate,
-            examModeCount,
-            examModeCorrectRate
+            todayLearningResponse.studyModeCount(),
+            todayLearningResponse.studyModeCorrectRate(),
+            todayLearningResponse.examModeCount(),
+            todayLearningResponse.examModeCorrectRate()
         );
     }
 }
